+++ apps/web/src/components/Home/OpenActions/Zora/ZoraNft.tsx
