+++ packages/helpers/generic/index.ts
@@ -6,12 +6,12 @@ export * from './functions/formatTime'
 export * from './functions/getCategoryName'
 export * from './functions/getChannelCoverPicture'
 export * from './functions/getCollectModule'
-export * from './functions/getCrossChainNftMetadata'
 export * from './functions/getFromAttributes'
 export * from './functions/getIsDispatcherEnabled'
 export * from './functions/getIsFeatureEnabled'
 export * from './functions/getIsSensitiveContent'
 export * from './functions/getMetadataCid'
+export * from './functions/getOpenActionNftMetadata'
 export * from './functions/getProfilePicture'
 export * from './functions/getPublicationMediaUrl'
 export * from './functions/getRandomProfilePicture'
